Fix/xvfb zombie (#357)
