+++ qsiprep/workflows/recon/mrtrix.py
@@ -10,7 +10,7 @@ import nipype.pipeline.engine as pe
 import nipype.interfaces.utility as niu
 from nipype.interfaces import afni
 from ...interfaces.bids import ReconDerivativesDataSink
-from ...interfaces.reports import ReconPeaksReport, ConnectivityReport
+from ...interfaces.reports import CLIReconPeaksReport, ConnectivityReport
 from qsiprep.interfaces.mrtrix import (
     EstimateFOD, SS3TEstimateFOD, MRTrixIngress, SS3TDwi2Response, GlobalTractography,
     MRTrixAtlasGraph, SIFT2, TckGen, MTNormalize)
@@ -150,7 +150,7 @@ A single-shell-optimized multi-tissue CSD was performed using MRtrix3Tissue
 (https://3Tissue.github.io), a fork of MRtrix3 (@mrtrix3)"""
 
     # Make a visual report of the model
-    plot_peaks = pe.Node(ReconPeaksReport(), name='plot_peaks')
+    plot_peaks = pe.Node(CLIReconPeaksReport(), name='plot_peaks')
     ds_report_peaks = pe.Node(
         ReconDerivativesDataSink(extension='.png',
                                  desc="wmFOD",
@@ -181,7 +181,7 @@ A single-shell-optimized multi-tissue CSD was performed using MRtrix3Tissue
         (inputnode, plot_peaks, [('dwi_ref', 'background_image'),
                                  ('odf_rois', 'odf_rois')]),
         (resample_mask, plot_peaks, [('out_file', 'mask_file')]),
-        (plot_peaks, ds_report_peaks, [('out_report', 'in_file')]),
+        (plot_peaks, ds_report_peaks, [('peak_report', 'in_file')]),
         (create_mif, estimate_response, [('mif_file', 'in_file')]),
         (estimate_response, outputnode, [('wm_file', 'wm_txt'),
                                          ('gm_file', 'gm_txt'),
